Start to add an assumptions system
The only properly implemented assumption as for now is divisibility. This allows us to execute the prob_collatz example, because we recognize that inside the if-block, x must be divisible by 2, meaning we can multiply with 1/2 without checking for divisibility in every state of the GF. It should be easy to extend this assumptions system for more use cases.